Update and reformat supported distro list
This list was off-putting to me as a user at first because it did not
include `centos7`, which made it seem like the project had been
abandoned long ago. However, upon browsing `rep.logentries.com`'s
directory listings, I learned there was indeed `centos7` support, it
just was not documented.

This change adds `centos7` to the list of supported distros.